% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection_functions.R
\name{cde_build_url}
\alias{cde_build_url}
\title{Construct the URL that will be used for an API call.}
\usage{
cde_build_url(query, options = NULL, page = "1", per_page = "100")
}
\arguments{
\item{query}{The path part of the URL that will be appended to the endpoint URL.}

\item{options}{A list of options to be included, like "year=2008".}

\item{page}{Which page of the query results will be retrieved.}

\item{per_page}{The maximum number of results that will be retrieved
with this query.}
}
\value{
A URL that will query the Crime Database when accessed.
}
\description{
This function stitches together the endpoint, API key, and various options
  to create the URL that will be accessed to make the query.
}
