% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection_functions.R
\name{cde_get_data}
\alias{cde_get_data}
\title{Retrieve data via the Crime Data API.}
\usage{
cde_get_data(query, options = NULL)
}
\arguments{
\item{query}{The path part of the URL that will be appended to the endpoint URL.}

\item{options}{A list of options to be included, like "year=2008".}
}
\value{
A data set containing the results of the query.
}
\description{
Uses \code{cde_build_url} to build a query URL, then runs the query until
  all results have been fetched.
}
