% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_functions.R
\name{cde_get_estimates}
\alias{cde_get_estimates}
\title{Retrieve NIBRS estimates of crime totals.}
\usage{
cde_get_estimates(state = "", fields = "")
}
\arguments{
\item{state}{Two-letter abbreviation of the state.}

\item{fields}{Limit the response to these specific fields.}
}
\value{
Data set returned from the API query.
}
\description{
Omitting the state will fetch national data.
}
\examples{
cde_get_estimates()
cde_get_estimates("FL")

x <- cde_get_estimates()
plot(x$year, x$homicide, type="b")
}
