% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_functions.R
\name{cde_get_ht_agencies}
\alias{cde_get_ht_agencies}
\title{Retrieve data on agencies reporting data on human trafficking to UCR.}
\usage{
cde_get_ht_agencies(state = "", agency_ori = "", agency_name = "",
  year = "")
}
\arguments{
\item{state}{Two-letter abbreviation of the state.}

\item{agency_ori}{The ORI of the agency to be queried.}

\item{agency_name}{Name of the agency (exact match only).}

\item{year}{Limit data to this year or year range.}
}
\value{
Data set returned from the API query.
}
\description{
Omitting the state will fetch national data, but I don't recommend it
  with this function - lots of data.
}
\examples{
cde_get_ht_agencies("FL")
}
