% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_functions.R
\name{cde_get_ht_states}
\alias{cde_get_ht_states}
\title{Retrieve data on human trafficking from the UCR.}
\usage{
cde_get_ht_states(state = "", year = "")
}
\arguments{
\item{state}{Two-letter abbreviation of the state.}

\item{year}{Limit data to this year or year range.}
}
\value{
Data set returned from the API query.
}
\description{
Omitting the state will fetch national data.
}
\examples{
cde_get_ht_states()
cde_get_ht_states("FL")
}
