% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_functions.R
\name{cde_get_participation_agencies}
\alias{cde_get_participation_agencies}
\title{Retrieve data on agencies that participate in UCR.}
\usage{
cde_get_participation_agencies(state = "", state_name = "", fields = "",
  year = "", agency_ori = "", agency_name = "", reported = "",
  reported_nibrs = "", months_reported = "", nibrs_months_reported = "")
}
\arguments{
\item{state}{Two-letter abbreviation of the state.}

\item{state_name}{Name of the state.}

\item{fields}{Limit the response to these specific fields.}

\item{year}{Limit data to this year or year range.}

\item{agency_ori}{The ORI of the agency to be queried.}

\item{agency_name}{Name of the agency (exact match only).}

\item{reported}{Whether the agency reported data.}

\item{reported_nibrs}{Whether the agency reported data for NIBRS.}

\item{months_reported}{Months when agency reported data.}

\item{nibrs_months_reported}{Months when agency reported data for NIBRS.}
}
\value{
Data set returned from the API query.
}
\description{
Omitting the state will fetch national data.
}
\details{
Note that I won't provide an example for this one because I don't
  know the agency names or ORIs required for a query.  Just included this
  as a function because it's one of the possible queries.
}
