% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection_functions.R
\name{cde_get_scope}
\alias{cde_get_scope}
\title{Determine whether the argument represents a state.}
\usage{
cde_get_scope(state)
}
\arguments{
\item{state}{A two-letter state abbreviation or an empty string.}
}
\value{
String that represents either national or state-level scope
  for a query URL.
}
\description{
The function returns either /national if it's not a state, or
  /states/<abbr> if it is a state.
}
