% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection_functions.R
\name{cde_make_ht_query}
\alias{cde_make_ht_query}
\title{Construct a query for human trafficking data from the UCR.}
\usage{
cde_make_ht_query(category, state = "", ori = "", agency_name = "",
  year = "")
}
\arguments{
\item{state}{Two-letter abbreviation of the state.}

\item{agency_name}{Name of the agency (exact match only).}

\item{year}{Limit data to this year or year range.}
}
\value{
Data set returned from the API query.
}
\description{
Omitting the state will fetch national data.
}
\details{
Note that you should use \code{cde_get_ht_states} or \code{cde_get_ht_agencies}
  instead of calling this function directly.
}
