% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection_functions.R
\name{cde_make_scoped_query}
\alias{cde_make_scoped_query}
\title{Fetch either national or state-level data.}
\usage{
cde_make_scoped_query(category, variable = "", variable_list = NULL,
  offense_list = NULL, state = "", offense = "", fields = "")
}
\arguments{
\item{category}{The category for the query, used in the base URL.}

\item{variable}{A variable to group results.}

\item{variable_list}{A list of allowed variable settings for the
function that called this one.}

\item{offense_list}{A list of allowed offense values for the function
that called this one.}

\item{state}{A two-letter state abbreviation or an empty string for a national
query.}

\item{offense}{Limits the results to matches for this offense.}

\item{fields}{Limits results to specific fields.}
}
\value{
A data set containing the results of the query.
}
\description{
Uses \code{cde_get_scope} to determine the scoped portion of the URL, then uses
\code{cde_get_data} to retrieve the data via the API.
}
