% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryNIBRS.R
\docType{package}
\name{queryNIBRS}
\alias{queryNIBRS}
\alias{queryNIBRS-package}
\title{queryNIBRS: A package for querying NIBRS crime data summaries.}
\description{
The data is retrieved via the FBI's Crime Data API.  Accessing
data will require the use of an API key that can be obtained
from \url{https://api.data.gov/signup/}.
}
\section{Warning}{

The API is considered to be in development, so some functions
may not retrieve data or work as advertised.
}

\seealso{
Crime Data Explorer: \url{https://crime-data-explorer.fr.cloud.gov/}

Crime Data API: \url{https://crime-data-explorer.fr.cloud.gov/api}
}
