% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection_functions.R
\name{set_ucr_api_key}
\alias{set_ucr_api_key}
\title{Set the Crime Data API key.}
\usage{
set_ucr_api_key(new_api_key)
}
\arguments{
\item{new_api_key}{The new API key value.}
}
\description{
You can get an API key from \url{https://api.data.gov/signup/}.
}
\seealso{
\url{https://api.data.gov/signup/}
}
